cirque pointing driver: Move fancy processing code
Inertial cursor is usable by any sensor with a lift status.
Extracted into quantum/pointing_device_gestures.[ch], to be extracted
into other get_report() functions if desired.
For example, one could mount a PMW3360 upside down with a clear
sheet over the lens, then use !(report.motion & 0x08) as Z and flick a
finger on said clear surface as a pointing device.

Circular scroll is specific to a circle trackpad reporting absolute
coordinates, Cirque is currently the only sensor of this type in QMK.
Tap could be generalized to multiple sensors, keeping it to lower level
for now since tap enable will be a register write for relative mode.
These two live in drivers/sensors/cirque_pinnacle_gestures.[ch].
