+++ quantum/pointing_device_drivers.c
@@ -97,281 +97,20 @@ const pointing_device_driver_t pointing_device_driver = {
 // clang-format on
 
 #elif defined(POINTING_DEVICE_DRIVER_cirque_pinnacle_i2c) || defined(POINTING_DEVICE_DRIVER_cirque_pinnacle_spi)
-#    ifndef CIRQUE_PINNACLE_TAPPING_TERM
-#        include "action.h"
-#        include "action_tapping.h"
-#        define CIRQUE_PINNACLE_TAPPING_TERM GET_TAPPING_TERM(KC_BTN1, &(keyrecord_t){})
-#    endif
-#    ifndef CIRQUE_PINNACLE_TOUCH_DEBOUNCE
-#        define CIRQUE_PINNACLE_TOUCH_DEBOUNCE (CIRQUE_PINNACLE_TAPPING_TERM * 8)
-#    endif
-
-typedef struct {
-    bool tap_enable;
-    bool circular_scroll_enable;
-    bool cursor_glide_enable;
-} cirque_pinnacle_features_t;
-
-static cirque_pinnacle_features_t features = {.tap_enable = true};
-
-typedef struct {
-    uint16_t timer;
-    bool     touchDown;
-} trackpad_tap_context_t;
-
-static trackpad_tap_context_t tap;
-
-static report_mouse_t trackpad_tap(report_mouse_t mouse_report, pinnacle_data_t touchData) {
-    if (touchData.touchDown != tap.touchDown) {
-        tap.touchDown = touchData.touchDown;
-        if (!touchData.zValue) {
-            if (timer_elapsed(tap.timer) < CIRQUE_PINNACLE_TAPPING_TERM && tap.timer != 0) {
-                mouse_report.buttons = pointing_device_handle_buttons(mouse_report.buttons, true, POINTING_DEVICE_BUTTON1);
-                pointing_device_set_report(mouse_report);
-                pointing_device_send();
-#    if TAP_CODE_DELAY > 0
-                wait_ms(TAP_CODE_DELAY);
-#    endif
-                mouse_report.buttons = pointing_device_handle_buttons(mouse_report.buttons, false, POINTING_DEVICE_BUTTON1);
-                pointing_device_set_report(mouse_report);
-                pointing_device_send();
-            }
-        }
-        tap.timer = timer_read();
-    }
-    if (timer_elapsed(tap.timer) > (CIRQUE_PINNACLE_TOUCH_DEBOUNCE)) {
-        tap.timer = 0;
-    }
-
-    return mouse_report;
-}
-
-// extern this function for now
-void cirque_pinnacle_enable_tap(bool enable) {
-    features.tap_enable = enable;
-}
-
-typedef enum {
-    SCROLL_UNINITIALIZED,
-    SCROLL_DETECTING,
-    SCROLL_VALID,
-    NOT_SCROLL,
-} circular_scroll_status_t;
-
-typedef struct {
-    int8_t v;
-    int8_t h;
-    bool   suppress_touch;
-} circular_scroll_t;
-
-typedef struct {
-    float                    mag;
-    int16_t                  x;
-    int16_t                  y;
-    uint16_t                 z;
-    circular_scroll_status_t state;
-    bool                     axis;
-    // settings
-    float   outer_ring_pct; // width of outer ring
-    float   movement_pct;   // amount of movement before triggering scroll validation
-    float   movement_ratio; // ratio of movement along perimeter / movement towards center
-    uint8_t wheel_clicks;   // how many clicks to report in a half circle
-} circular_scroll_context_t;
-
-static circular_scroll_context_t scroll = {.outer_ring_pct = 33, .movement_pct = 6, .movement_ratio = 1.2, .wheel_clicks = 9};
-
-static circular_scroll_t circular_scroll(pinnacle_data_t touchData) {
-    circular_scroll_t report = {0, 0, false};
-    int16_t           x, y;
-    uint16_t          center = cirque_pinnacle_get_scale() / 2;
-    int8_t            wheel_clicks;
-    float             mag, ang, scalar_projection, scalar_rejection, parallel_movement, perpendicular_movement;
-    int32_t           dot, det;
-
-    if (touchData.zValue) {
-        // place origin at center of trackpad
-        x = (int16_t)(touchData.xValue - center);
-        y = (int16_t)(touchData.yValue - center);
-
-        // check if first touch
-        if (!scroll.z) {
-            report.suppress_touch = false;
-            // check if touch falls within outer ring
-            mag = hypotf(x, y);
-            if (mag / center >= (100.0 - scroll.outer_ring_pct) / 100.0) {
-                scroll.state = SCROLL_DETECTING;
-                scroll.x     = x;
-                scroll.y     = y;
-                scroll.mag   = mag;
-                // decide scroll axis:
-                //   vertical if started from righ half
-                //   horizontal if started from left half
-                // reverse for left hand? TODO?
-#    if defined(POINTING_DEVICE_ROTATION_90)
-                scroll.axis = y < 0;
-#    elif defined(POINTING_DEVICE_ROTATION_180)
-                scroll.axis = x > 0;
-#    elif defined(POINTING_DEVICE_ROTATION_270)
-                scroll.axis = y > 0;
-#    else
-                scroll.axis = x < 0;
-#    endif
-            }
-        } else if (scroll.state == SCROLL_DETECTING) {
-            report.suppress_touch = true;
-            // already detecting scroll, check movement from touchdown location
-            mag = hypotf(x - scroll.x, y - scroll.y);
-            if (mag >= scroll.movement_pct / 100.0 * center) {
-                // check ratio of movement towards center vs. along perimeter
-                // this distinguishes circular scroll from swipes that start from edge of trackpad
-                dot                    = (int32_t)scroll.x * (int32_t)x + (int32_t)scroll.y * (int32_t)y;
-                det                    = (int32_t)scroll.x * (int32_t)y - (int32_t)scroll.y * (int32_t)x;
-                scalar_projection      = dot / scroll.mag;
-                scalar_rejection       = det / scroll.mag;
-                parallel_movement      = fabs(scroll.mag - fabs(scalar_projection));
-                perpendicular_movement = fabs(scalar_rejection);
-                if (parallel_movement * scroll.movement_ratio > perpendicular_movement) {
-                    // not a scroll, release coordinates
-                    report.suppress_touch = false;
-                    scroll.state          = NOT_SCROLL;
-                } else {
-                    // scroll detected
-                    scroll.state = SCROLL_VALID;
-                }
-            }
-        }
-        if (scroll.state == SCROLL_VALID) {
-            report.suppress_touch = true;
-            dot                   = (int32_t)scroll.x * (int32_t)x + (int32_t)scroll.y * (int32_t)y;
-            det                   = (int32_t)scroll.x * (int32_t)y - (int32_t)scroll.y * (int32_t)x;
-            ang                   = atan2f(det, dot);
-            wheel_clicks          = roundf(ang * (float)scroll.wheel_clicks / M_PI);
-            if (wheel_clicks >= 1 || wheel_clicks <= -1) {
-                if (scroll.axis == 0) {
-                    report.v = -wheel_clicks;
-                } else {
-                    report.h = wheel_clicks;
-                }
-                scroll.x = x;
-                scroll.y = y;
-            }
-        }
-    }
-
-    scroll.z = touchData.zValue;
-    if (!scroll.z) scroll.state = SCROLL_UNINITIALIZED;
-
-    return report;
-}
-
-// extern this for now
-void cirque_pinnacle_enable_circular_scroll(bool enable) {
-    features.circular_scroll_enable = enable;
-}
-
-// To set a trackpad exclusively as scroll wheel: outer_ring_pct = 100, movement_pct = 0, movement_ratio = 0
-void cirque_pinnacle_configure_circular_scroll(float outer_ring_pct, float movement_pct, float movement_ratio, uint8_t wheel_clicks) {
-    scroll.outer_ring_pct = outer_ring_pct;
-    scroll.movement_pct   = movement_pct;
-    scroll.movement_ratio = movement_ratio;
-    scroll.wheel_clicks   = wheel_clicks;
-}
-
-typedef struct {
-    mouse_xy_report_t dx;
-    mouse_xy_report_t dy;
-    bool              valid;
-} cursor_glide_t;
-
-typedef struct {
-    float             trigger_pct;
-    float             coef;
-    float             v0;
-    int16_t           x;
-    int16_t           y;
-    uint16_t          z;
-    uint16_t          timer;
-    uint16_t          interval;
-    uint16_t          counter;
-    mouse_xy_report_t dx0;
-    mouse_xy_report_t dy0;
-} cursor_glide_context_t;
-
-static cursor_glide_context_t glide = {.trigger_pct = 2};
-
-static cursor_glide_t cursor_glide(void) {
-    cursor_glide_t report;
-    float          p;
-    int16_t        x, y;
-
-    glide.counter++;
-    // calculate current position
-    p            = glide.v0 * glide.counter - glide.coef * glide.counter * glide.counter / 2;
-    x            = (int16_t)(p * glide.dx0 / glide.v0);
-    y            = (int16_t)(p * glide.dy0 / glide.v0);
-    report.dx    = (mouse_xy_report_t)(x - glide.x);
-    report.dy    = (mouse_xy_report_t)(y - glide.y);
-    report.valid = true;
-    if (report.dx <= 1 && report.dx >= -1 && report.dy <= 1 && report.dy >= -1) {
-        glide.dx0 = 0;
-        glide.dy0 = 0;
-    }
-    glide.x     = x;
-    glide.y     = y;
-    glide.timer = timer_read();
-
-    return report;
-}
-
-static cursor_glide_t cursor_glide_check(void) {
-    cursor_glide_t invalid_report = {0, 0, false};
-    if (glide.z || (glide.dx0 == 0 && glide.dy0 == 0) || timer_elapsed(glide.timer) < glide.interval) {
-        return invalid_report;
-    } else {
-        return cursor_glide();
-    }
-}
-
-static cursor_glide_t cursor_glide_start(void) {
-    cursor_glide_t invalid_report = {0, 0, false};
-
-    glide.coef     = 0.4; // good enough default
-    glide.interval = 10;  // hardcode for 100sps
-    glide.timer    = timer_read();
-    glide.counter  = 0;
-    glide.v0       = (glide.dx0 == 0 && glide.dy0 == 0) ? 0.0 : hypotf(glide.dx0, glide.dy0); // skip trigonometry if not needed
-    glide.x        = 0;
-    glide.y        = 0;
-    glide.z        = 0;
-
-    if (glide.v0 < glide.trigger_pct * cirque_pinnacle_get_scale() / 100) {
-        // not enough velocity to be worth gliding, abort
-        glide.dx0 = 0;
-        glide.dy0 = 0;
-        return invalid_report;
-    }
-
-    return cursor_glide();
-}
-
-static void cursor_glide_update(mouse_xy_report_t dx, mouse_xy_report_t dy, uint16_t z) {
-    glide.dx0 = dx;
-    glide.dy0 = dy;
-    glide.z   = z;
-}
+static bool                   cursor_glide_enable;
+static cursor_glide_context_t glide = {.coef = 0.4 /* good default friction coef */, .interval = 10 /* 100sps */};
 
 // extern this for now
 void cirque_pinnacle_enable_cursor_glide(bool enable) {
-    features.cursor_glide_enable = enable;
+    cursor_glide_enable = enable;
 }
 
-void cirque_pinnacle_configure_cursor_glide(float trigger_pct) {
-    glide.trigger_pct = trigger_pct;
+void cirque_pinnacle_configure_cursor_glide(float trigger_inches) {
+    glide.trigger_px = roundf((float)trigger_inches * (float)CIRQUE_PINNACLE_DIAMETER_MM / 25.4f);
 }
 
 report_mouse_t cirque_pinnacle_get_report(report_mouse_t mouse_report) {
     pinnacle_data_t          touchData = cirque_pinnacle_read_data();
-    circular_scroll_t        scroll_report;
     cursor_glide_t           glide_report;
     mouse_xy_report_t        report_x = 0, report_y = 0;
     static mouse_xy_report_t x = 0, y = 0;
@@ -380,12 +119,12 @@ report_mouse_t cirque_pinnacle_get_report(report_mouse_t mouse_report) {
 #        error Cirque Pinnacle with relative mode not implemented yet.
 #    endif
 
-    if (features.cursor_glide_enable) {
-        glide_report = cursor_glide_check();
+    if (cursor_glide_enable) {
+        glide_report = cursor_glide_check(&glide);
     }
 
     if (!touchData.valid) {
-        if (features.cursor_glide_enable && glide_report.valid) {
+        if (cursor_glide_enable && glide_report.valid) {
             report_x = glide_report.dx;
             report_y = glide_report.dy;
             goto mouse_report_update;
@@ -402,29 +141,23 @@ report_mouse_t cirque_pinnacle_get_report(report_mouse_t mouse_report) {
     // Scale coordinates to arbitrary X, Y resolution
     cirque_pinnacle_scale_data(&touchData, cirque_pinnacle_get_scale(), cirque_pinnacle_get_scale());
 
-    if (features.circular_scroll_enable) {
-        scroll_report  = circular_scroll(touchData);
-        mouse_report.v = scroll_report.v;
-        mouse_report.h = scroll_report.h;
-        if (scroll_report.suppress_touch) {
-            goto mouse_report_update;
+    if (!cirque_pinnacle_gestures(&mouse_report, touchData)) {
+        if (x && y && touchData.xValue && touchData.yValue) {
+            report_x = (mouse_xy_report_t)(touchData.xValue - x);
+            report_y = (mouse_xy_report_t)(touchData.yValue - y);
         }
-    }
-    if (x && y && touchData.xValue && touchData.yValue) {
-        report_x = (mouse_xy_report_t)(touchData.xValue - x);
-        report_y = (mouse_xy_report_t)(touchData.yValue - y);
-    }
-    x = touchData.xValue;
-    y = touchData.yValue;
+        x = touchData.xValue;
+        y = touchData.yValue;
 
-    if (features.cursor_glide_enable) {
-        if (touchData.touchDown) {
-            cursor_glide_update(report_x, report_y, touchData.zValue);
-        } else if (!glide_report.valid) {
-            glide_report = cursor_glide_start();
-            if (glide_report.valid) {
-                report_x = glide_report.dx;
-                report_y = glide_report.dy;
+        if (cursor_glide_enable) {
+            if (touchData.touchDown) {
+                cursor_glide_update(&glide, report_x, report_y, touchData.zValue);
+            } else if (!glide_report.valid) {
+                glide_report = cursor_glide_start(&glide);
+                if (glide_report.valid) {
+                    report_x = glide_report.dx;
+                    report_y = glide_report.dy;
+                }
             }
         }
     }
@@ -432,10 +165,6 @@ report_mouse_t cirque_pinnacle_get_report(report_mouse_t mouse_report) {
 mouse_report_update:
     mouse_report.x = report_x;
     mouse_report.y = report_y;
-
-    if (features.tap_enable) {
-        mouse_report = trackpad_tap(mouse_report, touchData);
-    }
 
     return mouse_report;
 }
